Move data-augmentation transformations into the new data_augmentation package (#566)
* Moving data-augmentation transformations into data_augmentation package

* Adding copyrights
